test: continue scanner coverage test split
